Local development with docker & docker-compose.
Now the developer doesn't need an local installation of LAMP or similar.
Also it's easier to debug and reproduce bugs, as the version of the runtime and DB is pinpointed.

Add two files to gitignore, that is mounted when the container start.
Add xdebug support, as well as debug run configuration for intellij idea.